--- src/main/java/com/sparta/lafesta/mail/config/MailConfig.java
@@ -1,38 +1,0 @@
-package com.sparta.lafesta.mail.config;
-
-import java.util.Properties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
-
-@Configuration
-public class MailConfig {
-
-  @Bean
-  public JavaMailSender javaMailService() {
-    JavaMailSenderImpl javaMailSender = new JavaMailSenderImpl();
-
-    javaMailSender.setHost("smtp.naver.com");
-    javaMailSender.setUsername("haeinj1215@naver.com");
-    javaMailSender.setPassword("비밀번호");
-
-    javaMailSender.setPort(465);
-
-    javaMailSender.setJavaMailProperties(getMailProperties());
-
-    return javaMailSender;
-  }
-
-  private Properties getMailProperties() {
-    Properties properties = new Properties();
-    properties.setProperty("mail.transport.protocol", "smtp");
-    properties.setProperty("mail.smtp.auth", "true");
-    properties.setProperty("mail.smtp.starttls.enable", "true");
-    properties.setProperty("mail.debug", "true");
-    properties.setProperty("mail.smtp.ssl.trust", "smtp.naver.com");
-    properties.setProperty("mail.smtp.ssl.enable", "true");
-    return properties;
-  }
-
-}
