--- src/main/java/com/sparta/lafesta/mail/service/MailService.java
@@ -1,15 +1,0 @@
-package com.sparta.lafesta.mail.service;
-
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
-import java.io.UnsupportedEncodingException;
-
-public interface MailService {
-
-  MimeMessage createMessage(String receiver)
-      throws MessagingException, UnsupportedEncodingException;
-
-  String createKey();
-
-  String sendMessage(String receiver) throws Exception;
-}
