+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/lafesta
 spring.datasource.username=root
-spring.datasource.password=haeinsparta
+spring.datasource.password=????
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.show_sql=true
@@ -20,23 +20,9 @@ spring.mail.properties.mail.smtp.connectiontimeout=5000
 spring.mail.properties.mail.smtp.timeout=5000
 spring.mail.properties.mail.smtp.writetimeout=5000
 spring.mail.auth-code-expiration-millis=1800000
-
 spring.thymeleaf.enabled=true
 spring.thymeleaf.prefix=classpath:/templates/
 spring.thymeleaf.suffix=.html
 spring.thymeleaf.cache=false
-
-kakao.client.id = ${KAKAO_REST_API}
-kakao.redirect.url = http://localhost:8080/api/users/kakao/callback
-
-spring.mail.host=${MAIL_HOST}
-spring.mail.port=${MAIL_PORT}
-spring.mail.username=${MAIL_USERNAME}
-spring.mail.password=${MAIL_PASSWORD}
-spring.mail.properties.mail.smtp.auth=true
-spring.mail.properties.mail.smtp.starttls.enable=true
-spring.mail.properties.mail.smtp.starttls.required=true
-spring.mail.properties.mail.smtp.connectiontimeout=5000
-spring.mail.properties.mail.smtp.timeout=5000
-spring.mail.properties.mail.smtp.writetimeout=5000
-spring.mail.auth-code-expiration-millis=1800000
+kakao.client.id=d6419e3db68a9f72405f8250aee4aaa8
+kakao.redirect.url=http://localhost:8080/api/users/kakao/callback
